pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "interview-helper"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.4.6",
    "deepgram-sdk>=3.0.0",
    "numpy>=1.26",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
